refactor(auth): move JwtService & IJwtService to jwt-core, create JwtRoleSwitchService, update controller and tests
