+++ itachallenge-auth/build.gradle
@@ -44,6 +44,7 @@ dependencies {
     implementation 'io.jsonwebtoken:jjwt-impl:0.12.6'
     implementation 'io.jsonwebtoken:jjwt-jackson:0.12.6'
     implementation 'org.projectlombok:lombok:1.18.30'
+    implementation project(':jwt-core')
     annotationProcessor 'org.projectlombok:lombok:1.18.30'
 
     testImplementation group: 'com.squareup.okhttp3', name: 'mockwebserver', version: '4.12.0'
@@ -57,6 +58,7 @@ dependencies {
     testImplementation 'org.testcontainers:junit-jupiter:1.17.6'
     testImplementation 'com.squareup.okhttp3:mockwebserver:4.9.3'
     testImplementation 'org.junit.platform:junit-platform-suite-engine:1.8.1'
+    testImplementation project(':jwt-core')
 
 }
