+++ itachallenge-auth/src/main/java/com/itachallenge/auth/service/IJwtService.java
@@ -1,13 +1,13 @@
-package com.itachallenge.auth.service;
-import io.jsonwebtoken.Claims;
-
-import javax.crypto.SecretKey;
-
-public interface IJwtService {
-    SecretKey getSigningKey();
-    String generateToken(String username, String role, String uuid);
-    void validateToken(String token);
-    Claims extractAllClaims(String token);
-    String switchRole(String token, String requestedRole);
-    String extractBearerToken(String authHeader);
-}
+//package com.itachallenge.auth.service;
+//import io.jsonwebtoken.Claims;
+//
+//import javax.crypto.SecretKey;
+//
+//public interface IJwtService {
+//    SecretKey getSigningKey();
+//    String generateToken(String username, String role, String uuid);
+//    void validateToken(String token);
+//    Claims extractAllClaims(String token);
+//    String switchRole(String token, String requestedRole);
+//    String extractBearerToken(String authHeader);
+//}
